List operations.
Cons, car, cdr, null?, emptylist.
Hope it is correct and I am not fooling myself via confirmation-bias.

However, I do not like the fact that I cannot test this at all.
(the test relative to lists in tests.scm do not work because I don't
know how to represent the expected value of lists, so ...)