Some formatting and other house-keeping.
Moved the cond handler into the interp.scm file, don't yet know if this
is final or not, I'll think about whether I want to keep it separated or
straight up include it as a sub function of the `value-of` procedure,
should skip having to pass the environement data structure everytime and
so on.

But I'm really not sure.