Letrec for myltiple variables.
Took me an embarassingly long time, sometimes I feel like my
understanding is nothing but a feeble house of cards that can be
destroyed by the merest whisper of confusion in the abstract landscape
that is my two-brain celled skull.

I am not bitter.

I will need to check the diff for this to not forget how I did this.

I am not bitter.